Instead of using entry meta, use `gv-revision` status
After consulting with David Smith, I decided that the most flexible storage model is to store a copy of the original entry, unserialized in the database. This will better allow for unlimited revisions and future functionality.

I plan on implementing the functionality added in #5 but this is just getting the transition started.